Add warning for dependency diff when model loading
