+++ tests/pyfunc/test_spark.py
@@ -162,7 +162,9 @@ def test_spark_udf(spark, model_path):
         code_path=[os.path.dirname(tests.__file__)],
     )
 
-    with mock.patch("mlflow.pyfunc._warn_dependency_requirement_mismatches") as mock_check_fn:
+    with mock.patch(
+        "mlflow.utils.requirements_utils.warn_dependency_requirement_mismatches"
+    ) as mock_check_fn:
         reloaded_pyfunc_model = mlflow.pyfunc.load_model(model_path)
         mock_check_fn.assert_called_once()
